changed use case registering, discord command registering  and new commands
