+++ app/app.go
@@ -10,9 +10,8 @@ import (
 )
 
 type Application struct {
-	Env     *Env
-	Db      *gorm.DB
-	Discord *discord.Service
+	Env *Env
+	Db  *gorm.DB
 }
 
 func App() Application {
@@ -21,15 +20,12 @@ func App() Application {
 	port, _ := strconv.Atoi(app.Env.DBPort)
 	app.Db = database.NewDatabase(app.Env.DBHost, app.Env.DBUser, app.Env.DBPass, app.Env.DBName, uint(port))
 
-	app.Discord = discord.CreateDiscord(app.Env.DiscordToken)
-
-	api.NewApiService(app.Db)
-
-	app.Discord.InitCommands(app.Db, app.Env.AppID)
-	err := app.Discord.OpenBot()
+	err := discord.OpenBot(app.Env.DiscordToken, app.Env.AppID, app.Db)
 	if err != nil {
 		panic(err)
 	}
+
+	api.NewApiService(app.Db)
 
 	// recordRepo := record.NewRecordRepository(app.Db)
 	// record := models.Record{Amount: 69, SubstanceID: 2, UserID: 1, Time: time.Now()}
