--- generic/models/discord.go
@@ -1,7 +1,0 @@
-package models
-
-import "github.com/bwmarrin/discordgo"
-
-type Discord struct {
-	Session *discordgo.Session
-}
